add attribute test_method to the class because this makes it more easy to generate advanced reports using the results class.
In my use-case, we extract a lot more information for out documentation from
the class or the test method. Having the test_method in the results this
enables us to access them easily in our self written report generator (a LaTeX
generator). Maybe one time the class can be exposed with a stable interface
which makes multiple generators possible. This may be a great benefit for
other use cases too.